// src/generated/schemas/memberRegisterDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * GREGORI API
 * GREGORI 종합 쇼핑몰 REST API 명세서
 * OpenAPI spec version: 1.0.0
 */

export interface MemberRegisterDto {
  /** @pattern ^[가-힣]{2,10}$ */
  name: string;
  email: string;
  /** @pattern ^(?=(.*[a-zA-Z].*){2,})(?=.*\d.*)(?=.*\W.*)[a-zA-Z0-9\S]{8,15}$ */
  password: string;
}